+++ modal/functions.py
@@ -46,6 +46,7 @@ from ._utils.function_utils import (
     _process_result,
     _stream_function_call_data,
     get_function_type,
+    get_include_source_mode,
     is_async,
 )
 from ._utils.grpc_utils import retry_transient_errors
@@ -369,21 +370,6 @@ OriginalReturnType = typing.TypeVar(
 )  # differs from return type if ReturnType is coroutine
 
 
-def get_include_source_mode(function_or_app_specific: Optional[str]) -> IncludeSourceMode:
-    if function_or_app_specific is not None:
-        valid_str_values = IncludeSourceValue.__args__
-        lower_case_input = function_or_app_specific.lower()
-        if lower_case_input not in valid_str_values:
-            valid_values_str = ", ".join(valid_str_values)
-            raise ValueError(
-                f"Invalid `include_source` value: {function_or_app_specific}. Use one of: {valid_values_str}"
-            )
-            # explicitly set in app/function
-        return IncludeSourceMode(lower_case_input)
-
-    return config.get("automount")
-
-
 class _Function(typing.Generic[P, ReturnType, OriginalReturnType], _Object, type_prefix="fu"):
     """Functions are the basic units of serverless execution on Modal.
 
@@ -531,7 +517,8 @@ class _Function(typing.Generic[P, ReturnType, OriginalReturnType], _Object, type
         else:
             # skip any mount introspection/logic inside containers, since the function
             # should already be hydrated
-            # TODO: maybe the entire constructor should be exited early if not local?
+            # TODO: maybe the entire from_args loader should be exited early if not local?
+            #  since it will be hydrated
             all_mounts = []
 
         retry_policy = _parse_retries(
